Bug 1687441 part 21 - Change Atomics VM functions to take index as size_t instead of int32_t. r=anba
Differential Revision: https://phabricator.services.mozilla.com/D102738

+++ js/src/jit/CacheIRCompiler.h
@@ -806,7 +806,7 @@ class MOZ_RAII CacheIRCompiler {
 
   bool emitDoubleIncDecResult(bool isInc, NumberOperandId inputId);
 
-  using AtomicsReadWriteModifyFn = int32_t (*)(TypedArrayObject*, int32_t,
+  using AtomicsReadWriteModifyFn = int32_t (*)(TypedArrayObject*, size_t,
                                                int32_t);
 
   [[nodiscard]] bool emitAtomicsReadModifyWriteResult(
